Clean up Multiple Mock strings!? Added mocks for My plot
