+++ lib/data/model/mock/MockString.dart
