+++ lib/ui/myplot/PlotList.dart
@@ -159,15 +159,4 @@ class PlotList extends StatelessWidget {
     );
   }
 
-  //FIXME: Only is built for show that this buttons are not functional yet
-  static void _showToast(BuildContext context) {
-    final String toastText = "Not Implemented Yet";
-    final String toastButtonText = "BACK";
-    final scaffold = Scaffold.of(context);
-    scaffold.showSnackBar(SnackBar(
-      content: Text(toastText),
-      action: SnackBarAction(
-          label: toastButtonText, onPressed: scaffold.hideCurrentSnackBar),
-    ));
-  }
 }
